[autorelease] Allow disabling conversion of output files to text.
Before this commit, it happens unconditionally even on files that are
not valid UTF-8, corrupting them. After this commit, it only happens on
files which are a sequence of 0x20..0x7f bytes (inclusive), and can
be disabled using the new option dictionary on `Application.run`.

Major version bump because the option dictionary is a breaking change.