[autorelease] Use ES modules to deliver static resources.
This is a stop-gap solution before WASI-Virt becomes usable.

Before this commit, it is not possible to use `yowasp-pack-resources`
to pack resources for nextpnr (~100 MB worst case) due to out of memory
errors in `JSON.stringify`. Also, binary files would get corrupted
since they would be read as UTF-8 with automatic replacement to U+FFFD.

After this commit, binary files are base64-encoded before embedding
into the module, and are decoded when the module is loaded.