Change `fetch` polyfill to be browser-compatible.
I did not realize `process?.x` throws a ReferenceError if `process`
is not defined, oops. Because of this, before this commit the polyfill
had to be removed by esbuild when bundling for the browser.

After this commit, the polyfil need not be removed at all, since it
becomes a no-op in the browser. It is still compatible with esbuild
if `--define:process=undefined` is used.