Switch to using tar as a pack format.
This is intended to improve performance (especially when used together
with @chialab/esbuild-plugin-meta-url plugin, but also generally in
the browser when using many files over the "large file" limit), as well
as simplify some things conceptually. As a side effect, the share root
is now inspectable with common Unix tools.

The generated resource.js file does require `nanotar` at runtime (and
does require the use of a bundler for the browser now), but this isn't
considered important since the old format was expressly designed for
bundling anyways.