[breaking-change] Calculate download progress correctly.
It turns out that `ReadableStream` is not (by itself) a way to compute
download progress: the `Content-Length` header is post-Zstd, causing
download progress to go above 100%, and our files are particularly in
need of being compressed, given the high ratios achievable.

This commit changes the interface to export total artifact size from
the resource module. Instead of `() => import('app-resources.js')`,
the import should now be used as simply `import('app-resources.js')`,
since the resource file itself is very small.